+++ lib/src/exception/validation_exception.dart
@@ -1,0 +1,8 @@
+class ValidationException implements Exception {
+  final String? message;
+
+  const ValidationException(this.message);
+
+  @override
+  String toString() => 'ValidationException(message: $message)';
+}