+++ lib/src/l10n/validators/l10n_validator.dart
@@ -1,0 +1,4 @@
+abstract class L10nValidator {
+  bool validate(String value);
+  String getError(String value);
+}